Work on longest path implementation more
This adds more to the longest path function implementation. It's based
on the networkx function's algorithm for computing this. But in porting
to rust it's still not quite ready yet. There are some edge cases where
the returned depth doesn't match expectations. This still needs to be
debugged. But it's better than what was there before (hard coded 0
return).