Add testing and ci infrastructure
Now that some of the funcationality is getting more involved (and still
isn't working fully yet) having some tests to verify the functionality
is important. This commit adds a python testing directory where we can
put tests on the python side. In addition to having a basic travis ci
config run the testing on new changes. We might add rust interface
testing in the future too, but for right now testing the python side
is good enough.