Add adj and adj_direction functions
This commit adds adj and adj_direction methods to the PyDAG class.
Coming from networkx adj(n) is the equivalent to adj[n].items() and will
also return a dict of nodes and edge weights.

As part of this change edge_weights are changed from an empty struct
(which was used as a placeholder) to a PyObject. This means you can set
the edge weight to any python object.