Add is_isomorphic function that node matches
This commit adds a new function to the retworkx module,
is_isomorphic_node_match(), which is used to check whether 2 graphs are
isomorphic and that the node data matches given a user supplied
function. The only existing isomorphic checking function only compared
the structure of the graph and did not check the data at all.