Add function to get all edges between 2 nodes
Prior to this patch there was a single get_edge_data function which used
the petgraph/daggy find_edge() function to get an edge index between 2
nodes (and then subsequently get the weight for that edge). However,
this only returns a single edge which is limiting for a multigraph like
what PyDAG provides. To address this limition this ommit adds a new
function get_all_edge_data() which will return a list of edge data for
all the edges found between the 2 specified nodes.