Add python 3.8 support
Now that python 3.8 has been released this commit adds the necessary
pieces to add python 3.8 support to retworkx. It updates the trove
classifiers to advertise the 3.8 support in the package metadata, adds
ci jobs to run tests on python 3.8, and updates the cibuildwheel version
to the latest which adds support for 3.8 wheels.