Implement a running cycle check
The previous patch implemented a cycle check on each edge addition (and
incorrectly on node addition functions) using:
petgraph::algo::is_cyclic_directed(). While this worked, it scans the
entire graph to look for cycles. This is not ideal for large graphs
because additions become increasingly slow. To avoid this problem this
commit adds a running cycle check to the edge adding operations based
heavily on what was implemented in the daggy project (previously used
as the graphing library). This grealy reduces the run time for addition
operations while still maintaining run time checks for cycles.