Add is_direct_acyclic_graph function
This commit adds a dual of the networkx function
is_direct_acyclic_graph() for retworkx's PyDAG class. While the previous
commit added run time checking to ensure that any added edge doesn't add
a cycle if an end user wants to rescan the full dag for cycles having a
function to enable that makes sense.