Add extra methods to PyDAG class
This commit adds 3 new methods to the PyDAG class to add equivalent
functionality to some networkx methods. The first two are in_edges and
out_edges. These functions are basically the same as adj_direction
except it's split by direction and also changes the output format to
be a list of tuples with node indexes and edge data. This way one can
reproduce/recreate the edges in the graph easily. The third method is
out_degree, which is the missing dual of in_degree but for the other
direction.