Add support for len() and new get_node_data() method
This commit adds 2 more method implementations to the PyDAG class. First
it adds support for calling len() on PyDAG objects which will return the
number of nodes in the graph. The second is a new method get_node_data()
which will return the object stored for a node at a given index.