Update documentation
This commit updates the documentation to reflect the current state of
the repo. First it adds a definition for __len__() which was missing
from the previous commit. Additionally, it fixes the class preamble to
not mention daggy anymore since that's not used. It also removes the bit
about indexes changing on removal since the use of StableGraph
internally has removed that limitation. It now only mentions that nodes
and edges are indexed with integers instead of using the data/weight.