Fix wheel builds (#10)
* Fix wheel builds on linux

The latest 0.1.0 release failed the wheel build. This was due to changes
in the cibuildwheel project for the 1.0.0 release. It now supports
building with both manylinux1 and manylinux2010. Manylinux2010 includes
a version of curl that no longer has the --proto option which is what
was used for downloading rustup as part of the build workflow. To
workaround this issue this commit switches to using wget which should do
the certificate checking without any extra arguments. At the same time
this bumps the cibuildwheel version to the latest.

* Add setuptools rust to appveyor wheel job

For whatever reason setuptools-rust wasn't getting installed
automatically as part of the wheel building process. This commit adds it
as an explicit pre-build step.