Fix issues with multigraphs
A couple of utility functions, mainly *_edges() and *_degree were not
dealing with multigraphs correctly. This is because all these functions
were mistakenly dealing with nodes instead of edges. This would mean
results were missing or were undercounting in the case when there were
multiple edges between 2 nodes. This commit fixes this by switching to
use petgraph's edges_directed() method instead of the
neighbors_directed().