Add support for pickling/deepcopy on PyDAG
This commit adds support for pickling and running copy.deepcopy() on
PyDAG objects. To do this 3 things are done, first it sets the module
for the class definition to be 'retworkx' instead of the default
builtins. Then it adds a __getstate__ method to return a dictionary
that can be used reconstruct the contents of the graph and __setstate__
which takes in that dictionary and builds a new graph StableDiGraph
object from that. It's worth noting post a deepcopy() the node indices
may not be the same if there were any removals in the original PyDAG()
object.