Fix longest_path function
This commit addresses some long standing issues in the longest_path
function. Since we previously only cared about the path length (ie
depth) the function worked as expected, but not because of the correct
reasons. This commit fixes the biggest underlying issue is that it would
not start the reverse path at the correct node. It would start the path
at the parent of the first node. The length function was working
correctly (as networkx does) because it was couting the equivalent of
the length of the shortened path. When the longest_path function was
fixed it started returning too large a value. The length function was
then updated to properly count the edges in the path (ie length - 1) and
return that value.

The now unused pairwise function is removed, that was only needed in the
networkx case where the edges are weighted and you want to factor that
into the longest_path_length returned. But, since that's outside the
scope of this project for now there is no reason to keep it around.