Make test_linear deterministic
The test_linear test case would some times fail randomly depending on
the returned order of the topological sort. Since the longest path could
end with either node 'f' or node 'g' and the path length would still be
valid. To avoid this the graph in the test case is changed so that node
'g' is a child of node 'c' instead of node 'e'. This makes the longest
path returned deterministic so there is no ambiguity in the test.