Don't add duplicates to successors or predecessors
There was a bug in the output from the successors and predecessors
methods for the PyDAG class with multigraphs. If there were multiple
edges to a node the output list would return the same node data twice.
This was unexpected because there is only one successor node not
multiple. To fix this issue this commit only adds a node to the output
of these methods once.