Make cycle checking opt-in
Currently the PyDAG class always checks on add_edge() calls that the new
edge would not create a cycle in the dag. However, this comes at a
performance cost, and sometimes depending on the size and complexity of
the dag this can be large. To avoid this performance penalty by default
this commit changes the PyDAG class to not do this checking by default.
It adds a new attribute check_cycle to the class, if this is set to True
then cycle checking is enabled on subsequent calls to add_edge().