Add check for cycles to check_cycle setter
This commit addresses a potential issue in the check_cycle setter if a
cycle was previously introduced prior to setting it True. This could
could undetected and cause unexpected problems down the road. To address
this possibility this adds a call to is_directed_acyclic_graph() to
verify that a cycle is not already present in the dag when enabling the
runtime checking.