Add PyDAG method to get node indexes
Currently the nodes() method returns a list of node data, but if for any
reason there is a need to get a list of all the indexes in use by the
PyDAG object the only option right now is to use a function like
topological sort which will return the sorted indexes. However, if
sorting isn't a requirement this is slower than needed. This commit adds
a method node_indexes() which will return a list of all node indexes.