Use Dijkstra's algorithm for ancestors/descendants
The previous implementation of the ancestors and descendants functions
were exceedingly inefficient. They iterated over each node in the graph
and then did a connectivity check between that node and the source or
target. For large graphs this could result in major bottleneck because
this is needlessly traversing the graph multiple times. This commit
reworks the internals of these functions to use Dijkstra's algorithm to
find paths from or to the specified node. This actually matches how
networkx implements these functions and should significantly improve
the runtime performance of the functions. The only tradeoff with this is
that results come back unordered now (instead of node creation ordered
as before).