Update release artifact jobs
Since the release of cibuildwheel we no longer need to install it from
source to get support for non-x86 architectures. At the same time the
sdist job was missing the install of setuptools-rust and building the
sdist was failing. This commit updates all the release artifact jobs to
use a consistent version of cibuildwheel and fix the sdist job.