Add missing pieces for Qiskit project standards
This commit adds the missing pieces from the repository to bring it more
inline with the rest of the projects under Qiskit. Mainly this means
adding templates for bugs, enhancements, and pull requests in addition
to adding a code of conduct and contributing guide. It also converts the
README file to be markdown instead of rst to be consistent with the rest
of Qiskit.