Remove sudo from pip calls in travis wheel jobs (#51)
The travis x86 linux wheel job is not working because when it calls sudo
pip the job tries to use a python 2 version of pip which no longer works
(nor does it have a cibuildwheel package on pypi that is compatible with
it). This commit removes the sudo usage to ensure we're always using
python3 pip.