Fix issue with cargo build on macOS
When developing on macOS attempting to just build the shared lib without
installing via setuptools-rust would lead to a linker error. This lead
those developing on macOS to often use cargo build to get debug syntax
and compiler errors in a good format and then switch to using pip for
testing that library actually could link and compile. However, this
issue is actually documented in the PyO3 README and it explains that
macOS needs additional linker flags to work. This commit adds the
necessary flags which should enable cargo build to fully compile on
macOS.