Upgrade PyO3 and option to enable check_cycle at init (#54)
This commit upgrades the version of PyO3 being used to the latest
release 0.9.0 which had some api changes that needed to be taken into
account. Additionally, the new version has improved docs which show how
to add arguments to the __new__() constructor definition with a default
value. So while the new construct is being updated to use the new class
initializer an optional argument is added.