Add compose method for combining two graph objects
This commit adds a compose() method to the PyDiGraph object and PyGraph
object. This method takes in another PyDiGraph or PyGraph object respectively
and a dictionary on where to add edges between the nodes. The keys are node
indicies in the first graph and the value is a tuple of the node indices in
the other graph and the weight to use for that edge.

This is a first implementation and is not that efficient since it
requires looping over the graph structure of the other graph to manually
reconstruct it in the original graph. Ideally in the future we should be
able to do this primarily as memory/vector operations and only adjusting
the graph structure on the boundaries. However, there is no support in
doing this via petgraph yet (see petgraph/petgraph#276) so for a first
pass this just iterates over the graph strucutre which should still be
significantly faster than doing it in python.

Related to #61