Revert "Rely on travis for macos wheel builds (#52)"
In an effort to get more reliable and compatible MacOS wheel builds
at release time this commit switches back to using azure pipelines
for the wheel jobs to get a faster and more up to date MacOS image.
This commit reverts the switch back to travis which was done to get
10.13 compatible wheels built by using the older MacOS travis env.
This is the first step, the next commit will start setting the
appropriate compiler flags to set the minimum compat level to an older
MacOS version.

This reverts commit 8cd11d5873a8a0037ae3bb20ff4dbe0d7131163f.