Explicitly set MACOSX_DEPLOYMENT_TARGET during MacOS jobs
In order to maximize compatibility for macOS wheels this commit sets the
MACOSX_DEPLOYMENT_TARGET env variable during the wheel building jobs to
explicity say we're building with compatibility for MacOS 10.9. Then in
order to verify that we're building during CI tests with the same
version this commit also updates the macOS test jobs to set the same
flag so we do not test with a different build than we publish at release
time.

Fixes #62