Fix docs build
The sphinx builds on rtd have been failing for some time. Lookingat the
logs this is because the m2r sphinx extension is not installed in the
build environment. This commit adds a docs requirements file so that rtd
knows to install m2r as part of the docs build.