Remove invalid flag for coverage job (#72)
The coverage job requires setting the -Zno-landing-pads flag, however on
the current version of rust nightly this flag has been removed (see:
rust-lang/rust#70175) and the job is no longer working. This commit works
around this by removing the flag to fix the job (see mozilla/grcov#427
for more details).