Add PyGraph class for undirected graphs (#68)
* Add PyGraph class for undirected graphs

This commit adds a new class to retworkx PyGraph which is used to build
an undirected graph. Right now it only supports basic operations for
adding/removing nodes and edges and some graph queries (like neighbors,
etc). Additional, algorithms will be added for this in the future.

* Add initial tests

This commit adds initial testing for the new PyGraph class covering
basic usage of nodes, edges, and adj.

* Update tests and coverage

* Fix lint

* Fix compiler warning

* Add docs

* Update docs