Bump dependency versions (#73)
There have been recent releases of both PyO3 and petgraph that add some
bugfixes and api changes. It doesn't look like this directly effects any
of the usage in retworkx, but it's best to keep it up-to-date
(especially for PyO3 because there is ongoing work to remove
specialization which will allow building with rust stable once it's finished and
then subsequently released).