Fix handling of holes in __setstate__
This commit fixes an issue with the handling of holes in node indexes in
__setstate__. The previous version of the logic was incorrectly
calculating how many temporary nodes to create to artificially increase
the node count to match the value in state. Additionally, the previous
functionality was incorrectly assuming a node index wouldn't be reused,
this isn't the case. So the logic was adjust to add all temporary nodes,
then add the node from the state, and then remove the temporary nodes.
This will have the desired effect of handling holes correctly and have
the node indexes after a pickle or deepcopy match the state before.