Add support for deepcopy/pickle to PyGraph (#74)
* Add support for deepcopy/pickle to PyGraoh

The PyGraph class was missing a __getstate__ and __setstate__ function
which were needed to enable support for pickling and deepcopy. This
commit adds the missing methods so PyGraph objects can be deepcopied.
One note on the test is that there aren't isomorphism functions that
support the pygraph type yet due to typing constraints a new trait will
need to be added for both PyGraph and PyDAG to leverage the existing
function. Until then the tests just assert that deepcopy() doesn't raise
and returns a PyGraph object.

* Fix rustfmt

* Fix handling of holes in __setstate__

This commit fixes an issue with the handling of holes in node indexes in
__setstate__. The previous version of the logic was incorrectly
calculating how many temporary nodes to create to artificially increase
the node count to match the value in state. Additionally, the previous
functionality was incorrectly assuming a node index wouldn't be reused,
this isn't the case. So the logic was adjust to add all temporary nodes,
then add the node from the state, and then remove the temporary nodes.
This will have the desired effect of handling holes correctly and have
the node indexes after a pickle or deepcopy match the state before.