Use into_pyarray instead of to_pyarray
Previously the new functions here were using .to_pyarray() which will
take the rust side array allocate a new numpy array on the python heap
and then copy the elements to that new numpy array. This can get
inefficient for large matrixes because it has to copy each element.
Instead we can leverage .into_pyarray() which creates a numpy array with
a pointer to the array in the rust heap (which is not at all a confusing
distinction for a 2 character difference). This commit makes the change
to ensure the conversion to numpy doesn't become a bottleneck.