Bump version strings
While the next release is probably a little while away enough new
features have landed since 0.3.4 that we can safely know that the next
release is 0.4.0. In preparation for that this bumps the version strings
in the setup.py and Cargo.toml, this is mostly to facilitate testing so
that we know when we're comparing a released version vs current master.