Switch to hashbrown hashmap implementation (#78)
The hashbrown library provides a drop-in replacement for the stdlib
HashMap implementation with improved performance. Since rust 1.36.0 the
stdlib implementation has been using the same algorithm as hasbrown
but hashbrown is still faster because it uses the AHash hashing
algorithm which is a tradeoff for speed vs hashing DOS resistance. Since
the use case for HashMaps in retworkx doesn't really need protection
against denial of service collisions picking performance makes sense
here. In the future if a use case arises we can always make the hashmap
lib configurable or add the use of the 'ahash-compile-time' feature which
increases the DOS-resistance of the algorithm.

The other potential speedup that hasbrown can provide in the future is
support for parallel iterators. Right now this commit doesn't use it
because adding and testing parallel iteration is a separate task, but
that can potentially provide a useful speedup for several functions
(including lexicographical topological sort) which rely on iterating
over hashmaps currently.