Add add_nodes_from and add_edges_from methods
This commit adds 2 methods to PyDAG and PyGraph, add_nodes_from() and
add_edges_from(). These methods take in lists of objects to add to the
graph to minimize the back and forth calling when adding many objects to
the graph at once. add_nodes_from() takes in a list of objects which
will be the data for the node, and add_nodes_from() takes in a list of
tuples with 2 node indexes for where the edge should be added and the
3rd tuple element is the data for the edge.