Adjust sort key to be consistent with expectations
The sorting we were using was different then the equivalent function in
networkx this is trying to match. Instead of doing a stable sort and
then reversing this could result in a different sorting order than what
someone migrating from the equivalent networkx function would expect.
This commit changes the sort key to be the
{number of nodes} - {number of neighbors} to address this. It also has
the advantage of speeding it up because we don't have to do the inplace
reverse.