Fix edge case with *_adjacency_matrix() functions (#80)
This commit fixes an edge case with the adjacency matrix functions
around holes in the node indexes. Previously the nodes were assumed to
be contiguous but in the case of node removals it was possible to have
holes in the list of indexes. This would result in a matrix that was
larger than expected because it was building it based on the maximum
number of nodes during the lifetime of the graph. To fix this issue a
new attribute is added to PyDAG and PyGraph to track if nodes have been
removed. If nodes have been removed a mapping of node_ids to compressed
indexes are built and as the edges are iterated over the compressed
index is used in the matrix. If no nodes are removed the mapping step
is skipped since it is unecessary.