Add edges from no data methods
In networkx there the add_edges_from() methods support either 2 or 3
tuples to optionally include a data payload, but not require it.
Implementing this in rust is a bit trickier since the function inputs
need to be of a known size at compile time so having a Vec of arbitrary
length tuples isn't really an option. In a pure rust program we'd
probably use an Option<PyObject> for the third tuple element, but that
doesn't translate to python. To avoid the overhead of having to loop
over lists in python that are using 2 tuples to add a None this commit
adds a second method add_edges_from_no_data that takes in a 2 tuple and
sets the data payload to None when the edge is added.