Calculate sort keys upfront
This commit reworks the sorting to calculate the keys upfront all at
once prior to doing the sorting and storing them in a hashmap. Then the
sort function is just based on a hashmap value lookup. This makes the
number of times we run sorting key function (which gets a count of
neighbors and subtracts the node count from it) less which speeds up the
run time.

Co-Authored-by: Matthew Treinish <mtreinish@kortar.org>