Fix docstrings and rework/simplify cycle checking logic
This commit fixes some typos in the docs for the new functions and
cleans them up slightly. Additionally, the logic and code around the
cycle checking for PyDAG has been reworked. The logic is simplified
slightly and the variable names are changed to make it clearer what is
happening in there. It also adds some comments to try and explain the
logic there a bit better.