Temporarily pin nightly version used for builds
In the most recent nightly build of cargo a regression slipped in around
handling of READMEs in published crates [1]. Unfortunately this breaks CI
builds because a 3rd level dependency (pyo3->inventory->ctor) triggers
this regression. The issue has been corrected in cargo and the next
nightly should fix this. However, to unblock things in the meantime this
commit pins the nightly version to a known working version. We can
simply revert this after the fix has been released.

[1] https://github.com/rust-lang/cargo/issues/8351