Error on invalid input index and update tests
This commit updates the *all_simple_paths functions so that an error is
raised if an input node index is not in the graph object. Previously
this wasn't be handled as an error case and an empty path list was
silently returned. At the same time the unittests are updated to test
this edge case and simplify the logic in the tests.