Handle edge case of large degrees in multigraphs
This commit fixes an edge case in the new function where we'd get
incorrect results if the degree of a node was greater than the number of
nodes in the graph. We previously would cause an overflow by subtracting
number of nodes - degree and the reverse ordering was broken because the
node with the large degree is now a larger number instead of a smaller
number. This fixes this issue by using the Reverse struct from std::cmp
[1] and making the sort key Reverse(degree). The hash map previously
used is kept in place for performance. A test is also added to test this
edge case.

[1] https://doc.rust-lang.org/std/cmp/struct.Reverse.html