Add astar shortest_path function for PyGraph and PyDAG (#84)
* Add astar shortest_path function for PyGraph

This commit adds a new function for finding the shortest path from a
node to a node that satisfies an arbitrary condition. This is built
using the astar algorithm from the upstream petgraph lib, however
because the data for the graphs being traversed are PyObjects we needed
to be able to handle python exceptions gracefully (instead of panicking)
which was not possible with the upstream astar function. So this
function has been forked into retworkx and modified to hand PyResult<>
wrapped returns from the callables.

* Split graph and dag impl and add tests and docs

This commit splits out the definition for graph and dag based astar
searches. Then tests and docs are added for both.

* Fix lint in tests

The hardest part of switching between rust and python is remembering to
use semicolons or not.

* Apply suggestions from code review

Co-authored-by: Lauren Capelluto <laurencapelluto@gmail.com>

* Fix duplicate typos

Co-authored-by: Lauren Capelluto <laurencapelluto@gmail.com>