Add all_simple_paths for dag and graph (#85)
* Add all_simple_paths for dag and graph

This commit adds 2 new functions dag_all_simple_paths() and
graph_all_simple_paths() to find all simple paths between 2 nodes in a
PyDAG and PyGraph object.

* Add min_depth kwarg and change defaults None

This commit adds a new kwarg to dag_all_simple_paths()
graph_all_simple_paths() min_depth which is used to specify a minimum
depth for the paths to include in the minimum depth. In addition the
default values for the kwargs on both method are set to None instead of
-1.

* Apply suggestions from code review

Co-authored-by: Lauren Capelluto <laurencapelluto@gmail.com>

* Error on invalid input index and update tests

This commit updates the *all_simple_paths functions so that an error is
raised if an input node index is not in the graph object. Previously
this wasn't be handled as an error case and an empty path list was
silently returned. At the same time the unittests are updated to test
this edge case and simplify the logic in the tests.

* Fix lint

* Fix missing '}' from rebase

The github web ui should be avoided, it's too easy to miss syntax errors
like this.

Co-authored-by: Lauren Capelluto <laurencapelluto@gmail.com>